Update configuration for routing templates
