+++ routing-typescript/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## 1.0.0
+## v1.1.0
+
+### Changed
+-   Updated dependencies to use latest packages.
+
+### Fixed
+-   Jest configuration allows default tests to pass.
+
+## v1.0.0
 
 ### Added
