+++ routing-typescript/template/__mocks__/svgMock.js
@@ -1,0 +1,8 @@
+/**
+ * SVG setup to mock react components transformed into images from SVGs
+ * Place this file in your project's root __mocks__ directory.
+ * https://github.com/kristerkari/react-native-svg-transformer#usage-with-jest
+ */
+
+ module.exports = 'SvgMock'
+ module.exports.ReactComponent = 'SvgMock'
