+++ routing-typescript/template/screens/home.tsx
@@ -16,7 +16,7 @@ import { Theme } from 'react-native-paper/lib/typescript/types';
 import Logo from '../assets/images/Logo.svg';
 import MatIcon from 'react-native-vector-icons/MaterialIcons';
 import { StackNavigationProp } from '@react-navigation/stack';
-import { RootStackParamList } from '../router';
+import { RootStackParamList } from '../navigation';
 
 const MenuIcon = wrapIcon({ IconClass: MatIcon, name: 'menu', flip: false });
 
@@ -129,7 +129,7 @@ const Home: React.FC<AppProps> = ({ navigation }): JSX.Element => {
                         Welcome to PX <H4 color={'primary'}>Blue</H4>.
                     </H4>
                     <Body1 style={defaultStyles.subtitle}>
-                        Edit <Body1 style={defaultStyles.bold}>pages/home.tsx</Body1> and save to reload.
+                        Edit <Body1 style={defaultStyles.bold}>screens/home.tsx</Body1> and save to reload.
                     </Body1>
                     <Divider style={defaultStyles.divider} />
                     <OpenURLButton title={'PX Blue Documentation'} url={'https://pxblue.github.io/'} />
