+++ routing-typescript/template/screens/pageOne.tsx
@@ -3,7 +3,7 @@ import { SafeAreaView, ScrollView, StyleSheet, ViewStyle } from 'react-native';
 import { EmptyState, Header, wrapIcon } from '@pxblue/react-native-components';
 import MatIcon from 'react-native-vector-icons/MaterialIcons';
 import { StackNavigationProp } from '@react-navigation/stack';
-import { RootStackParamList } from '../router';
+import { RootStackParamList } from '../navigation';
 const Event = wrapIcon({ IconClass: MatIcon, name: 'event', flip: false });
 const MenuIcon = wrapIcon({ IconClass: MatIcon, name: 'menu', flip: false });
 
@@ -23,16 +23,16 @@ const styles = (): StyleSheet.NamedStyles<{
     });
 
 type AppProps = {
-    navigation: StackNavigationProp<RootStackParamList, 'PageTwo'>;
+    navigation: StackNavigationProp<RootStackParamList, 'PageOne'>;
 };
 
-const PageTwo: React.FC<AppProps> = ({ navigation }): JSX.Element => {
+const PageOne: React.FC<AppProps> = ({ navigation }): JSX.Element => {
     const defaultStyles = styles();
 
     return (
         <>
             <Header
-                title={'Page Two'}
+                title={'Page One'}
                 navigation={{
                     icon: MenuIcon,
                     onPress: (): void => {
@@ -53,4 +53,4 @@ const PageTwo: React.FC<AppProps> = ({ navigation }): JSX.Element => {
     );
 };
 
-export default PageTwo;
+export default PageOne;
