+++ routing/dependencies.json
@@ -1,22 +1,24 @@
 {
     "dependencies": [
-        "@pxblue/react-native-themes@^5.0.0",
+        "@pxblue/react-native-themes@^5.1.0",
         "@pxblue/colors@^3.0.0",
-        "@pxblue/react-native-components@^4.1.0",
-        "react-native-modal@^11.6.1",
-        "react-native-paper@^4.2.0",
-        "@pxblue/icons-svg@^1.1.1",
-        "react-native-vector-icons@^7.1.0",
+        "@pxblue/react-native-components@^5.0.0",
+        "react-native-modal@^11.7.0",
+        "react-native-paper@^4.7.2",
+        "@pxblue/icons-svg@^1.2.0",
+        "react-native-vector-icons@^8.0.0",
         "react-native-svg@^12.1.0",
         "react-native-svg-transformer@^0.14.3",
-        "react-native-safe-area-context@^3.1.9",
+        "react-native-safe-area-context@^3.2.0",
         "@react-native-community/masked-view@^0.1.10",
         "@react-navigation/drawer@^5.12.3",
-        "@react-navigation/native@^5.9.2",
-        "@react-navigation/stack@^5.14.2",
-        "react-native-gesture-handler@^1.10.1",
-        "react-native-reanimated@^1.13.2",
-        "react-native-screens@^2.17.1"
+        "@react-navigation/native@^5.9.3",
+        "@react-navigation/stack@^5.14.3",
+        "react-native-gesture-handler@^1.10.3",
+        "react-native-reanimated@^2.0.1",
+        "react-native-screens@^3.0.0"
     ],
-    "devDependencies": ["jest", "react-test-renderer", "@types/react-native-vector-icons"]
+    "devDependencies": [
+        "jest-transform-stub"
+    ]
 }
