+++ routing/template/metro.config.js
@@ -17,7 +17,7 @@ module.exports = (async () => {
             getTransformOptions: async () => ({
                 transform: {
                     experimentalImportSupport: false,
-                    inlineRequires: false,
+                    inlineRequires: true,
                 },
             }),
         },
