+++ routing/template/navigation/navigation-drawer.js
@@ -1,16 +1,15 @@
-import { Drawer, DrawerBody, DrawerHeader, DrawerNavGroup, NavItem, wrapIcon } from '@pxblue/react-native-components';
+import { Drawer, DrawerBody, DrawerHeader, DrawerNavGroup, wrapIcon } from '@pxblue/react-native-components';
 import React, { useState, useCallback } from 'react';
 import MatIcon from 'react-native-vector-icons/MaterialIcons';
 import { IconButton } from 'react-native-paper';
 import * as Colors from '@pxblue/colors';
-import { StackNavigationProp } from '@react-navigation/stack';
-import { RootStackParamList } from './index';
 
+const Menu = wrapIcon({ IconClass: MatIcon, name: 'menu', flip: false });
 const Home = wrapIcon({ IconClass: MatIcon, name: 'home', flip: false });
 const LooksOne = wrapIcon({ IconClass: MatIcon, name: 'looks-one', flip: false });
 const LooksTwo = wrapIcon({ IconClass: MatIcon, name: 'looks-two', flip: false });
 
-export const navGroupItems: NavItem[] = [
+export const navGroupItems = [
     {
         title: 'Home Page',
         itemID: 'Home',
@@ -28,11 +27,7 @@ export const navGroupItems: NavItem[] = [
     },
 ];
 
-export type NavDrawerProps = {
-    navigation: StackNavigationProp<RootStackParamList, 'NavigationDrawer'>;
-};
-
-export const NavigationDrawer: React.FC<NavDrawerProps> = ({ navigation }) => {
+export const NavigationDrawer = ({ navigation }) => {
     const [selected, setSelected] = useState('Home');
     const selectItem = useCallback(
         (id) => {
@@ -43,21 +38,17 @@ export const NavigationDrawer: React.FC<NavDrawerProps> = ({ navigation }) => {
     );
 
     return (
-        <Drawer activeItem={selected} onItemSelect={(id: string): void => selectItem(id)}>
+        <Drawer activeItem={selected} onItemSelect={(id) => selectItem(id)}>
             <DrawerHeader
                 title={'PX Blue'}
                 subtitle={'React Native Project'}
                 fontColor={Colors.white[50]}
-                icon={
-                    <IconButton
-                        icon="menu"
-                        size={24}
-                        color={Colors.white[50]}
-                        onPress={(): void => {
-                            navigation.closeDrawer();
-                        }}
-                    />
-                }
+                icon={{
+                    icon: Menu,
+                    onPress: (): void => {
+                        navigation.closeDrawer();
+                    },
+                }}
             />
             <DrawerBody>
                 <DrawerNavGroup items={navGroupItems} hidePadding={false} />
